kafka_producer/requirements.txt
kafka-python==2.0.2
schedule==1.1.0
faker==13.3.4
